+++ Druger/Druger/DOJO/DRUG_FACTORY.swift
@@ -10,7 +10,7 @@ class DRUG_FACTORY: CamembertModel {
     //名字
     var NAME:TEXT = ""
     //价格
-    var PRICE:INTEGER = 0
+    var PRICE:REAL = 0
     //数量
     var COUNT:INTEGER = 0
     //是否可以被显示
@@ -18,7 +18,7 @@ class DRUG_FACTORY: CamembertModel {
     //制毒人员
     var WORKER_NAME:TEXT = ""
     //制度人员价格
-    var WORKER_PRICE:INTEGER = 0
+    var WORKER_PRICE:REAL = 0
     //制毒人员制毒速度
     var WORKER_SPEED:INTEGER = 0
     //制毒人员人数
@@ -36,7 +36,7 @@ class DRUG_FACTORY: CamembertModel {
         super.init()
     }
     
-    init(id:Int, NAME:String, PRICE:Int64, WORKER_NAME:String, WORKER_PRICE:Int64, WORKER_SPEED:Int64, WORKER_MAX:Int64, DESCRIBE:TEXT) {
+    init(id:Int, NAME:String, PRICE:Float, WORKER_NAME:String, WORKER_PRICE:Float, WORKER_SPEED:Int64, WORKER_MAX:Int64, DESCRIBE:TEXT) {
         super.init(id: id)
         self.NAME = NAME
         self.PRICE = PRICE
